+++ client/src/lib.rs
@@ -34,7 +34,10 @@ use common::{
     outcome::Outcome,
     recipe::RecipeBook,
     resources::{DeltaTime, PlayerEntity, TimeOfDay},
-    terrain::{block::Block, neighbors, BiomeKind, SitesKind, TerrainChunk, TerrainChunkSize},
+    terrain::{
+        block::Block, map::MapConfig, neighbors, BiomeKind, SitesKind, TerrainChunk,
+        TerrainChunkSize,
+    },
     trade::{PendingTrade, SitePrices, TradeAction, TradeId, TradeResult},
     uid::{Uid, UidAllocator},
     vol::RectVolSize,
@@ -112,17 +115,20 @@ pub struct WorldData {
     /// map data (e.g. with shadow map data or river data), but at present
     /// we opt not to do this.
     ///
-    /// The second element of the tuple is the world size (as a 2D grid,
-    /// in chunks), and the third element holds the minimum height for any land
-    /// chunk (i.e. the sea level) in its x coordinate, and the maximum land
-    /// height above this height (i.e. the max height) in its y coordinate.
-    map: (Arc<DynamicImage>, Vec2<u16>, Vec2<f32>),
+    /// The first two elements of the tuple are the regular and topographic maps
+    /// respectively. The third element of the tuple is the world size (as a 2D
+    /// grid, in chunks), and the fourth element holds the minimum height for
+    /// any land chunk (i.e. the sea level) in its x coordinate, and the maximum
+    /// land height above this height (i.e. the max height) in its y coordinate.
+    map: (Vec<Arc<DynamicImage>>, Vec2<u16>, Vec2<f32>),
 }
 
 impl WorldData {
     pub fn chunk_size(&self) -> Vec2<u16> { self.map.1 }
 
-    pub fn map_image(&self) -> &Arc<DynamicImage> { &self.map.0 }
+    pub fn map_layers(&self) -> &Vec<Arc<DynamicImage>> { &self.map.0 }
+
+    pub fn map_image(&self) -> &Arc<DynamicImage> { &self.map.0[0] }
 
     pub fn min_chunk_alt(&self) -> f32 { self.map.2.x }
 
@@ -324,6 +330,9 @@ impl Client {
 
                 // Redraw map (with shadows this time).
                 let mut world_map_rgba = vec![0u32; rgba.size().product() as usize];
+                let mut world_map_political = vec![0u32; rgba.size().product() as usize];
+                let mut world_map_rgba_half_alpha = vec![0u32; rgba.size().product() as usize];
+                let mut world_map_topo = vec![0u32; rgba.size().product() as usize];
                 let mut map_config = common::terrain::map::MapConfig::orthographic(
                     map_size_lg,
                     core::ops::RangeInclusive::new(0.0, max_height),
@@ -336,49 +345,140 @@ impl Client {
                         && pos.y < map_size.y as i32
                 };
                 ping_stream.send(PingMsg::Ping)?;
+                fn sample_pos(
+                    map_config: &MapConfig,
+                    pos: Vec2<i32>,
+                    alt: &Grid<u32>,
+                    rgba: &Grid<u32>,
+                    map_size: &Vec2<u16>,
+                    map_size_lg: &common::terrain::MapSizeLg,
+                    max_height: f32,
+                ) -> common::terrain::map::MapSample {
+                    let rescale_height = |h: f32| h / max_height;
+                    let scale_height_big = |h: u32| (h >> 3) as f32 / 8191.0 * max_height;
+                    let bounds_check = |pos: Vec2<i32>| {
+                        pos.reduce_partial_min() >= 0
+                            && pos.x < map_size.x as i32
+                            && pos.y < map_size.y as i32
+                    };
+                    let MapConfig {
+                        gain,
+                        is_contours,
+                        is_height_map,
+                        is_political,
+                        is_roads,
+                        rgba_alpha,
+                        ..
+                    } = *map_config;
+                    let mut is_contour_line = false;
+                    let mut is_border = false;
+                    let (rgba, alt, downhill_wpos) = if bounds_check(pos) {
+                        let posi = pos.y as usize * map_size.x as usize + pos.x as usize;
+                        let [r, g, b, a] = rgba[pos].to_le_bytes();
+                        let is_water = r == 0 && b > 102 && g < 77;
+                        let alti = alt[pos];
+                        // Compute contours (chunks are assigned in the river code below)
+                        let altj = rescale_height(scale_height_big(alti));
+                        let contour_interval = 150.0;
+                        let chunk_contour = (altj * gain / contour_interval) as u32;
+
+                        // Compute downhill.
+                        let downhill = {
+                            let mut best = -1;
+                            let mut besth = alti;
+                            for nposi in neighbors(*map_size_lg, posi) {
+                                let nbh = alt.raw()[nposi];
+                                let nalt = rescale_height(scale_height_big(nbh));
+                                let nchunk_contour = (nalt * gain / contour_interval) as u32;
+                                if !is_contour_line && chunk_contour > nchunk_contour {
+                                    is_contour_line = true;
+                                }
+                                let [nr, ng, nb, _na] = rgba.raw()[nposi].to_le_bytes();
+                                let n_is_water = nr == 0 && nb > 102 && ng < 77;
+
+                                if !is_border && is_political && is_water && !n_is_water {
+                                    is_border = true;
+                                }
+
+                                if nbh < besth {
+                                    besth = nbh;
+                                    best = nposi as isize;
+                                }
+                            }
+                            best
+                        };
+                        let downhill_wpos = if downhill < 0 {
+                            None
+                        } else {
+                            Some(
+                                Vec2::new(
+                                    (downhill as usize % map_size.x as usize) as i32,
+                                    (downhill as usize / map_size.x as usize) as i32,
+                                ) * TerrainChunkSize::RECT_SIZE.map(|e| e as i32),
+                            )
+                        };
+                        (Rgba::new(r, g, b, a), alti, downhill_wpos)
+                    } else {
+                        (Rgba::zero(), 0, None)
+                    };
+                    let alt = f64::from(rescale_height(scale_height_big(alt)));
+                    let wpos = pos * TerrainChunkSize::RECT_SIZE.map(|e| e as i32);
+                    let downhill_wpos = downhill_wpos
+                        .unwrap_or(wpos + TerrainChunkSize::RECT_SIZE.map(|e| e as i32));
+                    let is_path = rgba.r == 0x37 && rgba.g == 0x29 && rgba.b == 0x23;
+                    let rgba = rgba.map(|e: u8| e as f64 / 255.0);
+                    let rgba = if is_height_map {
+                        if is_path {
+                            // Path color is Rgb::new(0x37, 0x29, 0x23)
+                            Rgba::new(0.9, 0.9, 0.63, 1.0)
+                        } else if rgba.r == 0.0 && rgba.b > 0.4 && rgba.g < 0.3 {
+                            // Water
+                            Rgba::new(0.23, 0.47, 0.53, 0.5)
+                        } else if is_contours && is_contour_line {
+                            // Color contour lines
+                            Rgba::new(0.15, 0.15, 0.15, 0.5)
+                        } else {
+                            // Color hill shading
+                            let lightness = (alt + 0.2).min(1.0) as f64;
+                            Rgba::new(lightness, 0.9 * lightness, 0.5 * lightness, 0.5)
+                        }
+                    } else if is_roads && is_path {
+                        Rgba::new(0.9, 0.9, 0.63, 1.0)
+                    } else if is_political {
+                        if is_path {
+                            Rgba::new(0.3, 0.3, 0.3, 1.0)
+                        } else if is_border {
+                            Rgba::new(0.0, 0.0, 0.0, 1.0)
+                        } else {
+                            Rgba::new(1.0, 0.9, 0.6, 1.0)
+                        }
+                    } else if is_contours && is_contour_line {
+                        Rgba::new(0.15, 0.15, 0.15, 0.8)
+                    } else {
+                        Rgba::new(rgba.r, rgba.g, rgba.b, rgba_alpha)
+                    }
+                    .map(|e| (e * 255.0) as u8);
+                    common::terrain::map::MapSample {
+                        rgba,
+                        alt,
+                        downhill_wpos,
+                        connections: None,
+                        is_path,
+                    }
+                }
+                map_config.is_shaded = true;
+                map_config.rgba_alpha = 1.0;
                 map_config.generate(
                     |pos| {
-                        let (rgba, alt, downhill_wpos) = if bounds_check(pos) {
-                            let posi = pos.y as usize * map_size.x as usize + pos.x as usize;
-                            let [r, g, b, a] = rgba[pos].to_le_bytes();
-                            let alti = alt[pos];
-                            // Compute downhill.
-                            let downhill = {
-                                let mut best = -1;
-                                let mut besth = alti;
-                                for nposi in neighbors(map_size_lg, posi) {
-                                    let nbh = alt.raw()[nposi];
-                                    if nbh < besth {
-                                        besth = nbh;
-                                        best = nposi as isize;
-                                    }
-                                }
-                                best
-                            };
-                            let downhill_wpos = if downhill < 0 {
-                                None
-                            } else {
-                                Some(
-                                    Vec2::new(
-                                        (downhill as usize % map_size.x as usize) as i32,
-                                        (downhill as usize / map_size.x as usize) as i32,
-                                    ) * TerrainChunkSize::RECT_SIZE.map(|e| e as i32),
-                                )
-                            };
-                            (Rgba::new(r, g, b, a), alti, downhill_wpos)
-                        } else {
-                            (Rgba::zero(), 0, None)
-                        };
-                        let wpos = pos * TerrainChunkSize::RECT_SIZE.map(|e| e as i32);
-                        let downhill_wpos = downhill_wpos
-                            .unwrap_or(wpos + TerrainChunkSize::RECT_SIZE.map(|e| e as i32));
-                        let alt = rescale_height(scale_height_big(alt));
-                        common::terrain::map::MapSample {
-                            rgb: Rgb::from(rgba),
-                            alt: f64::from(alt),
-                            downhill_wpos,
-                            connections: None,
-                        }
+                        sample_pos(
+                            &map_config,
+                            pos,
+                            &alt,
+                            &rgba,
+                            &map_size,
+                            &map_size_lg,
+                            max_height,
+                        )
                     },
                     |wpos| {
                         let pos = wpos.map2(TerrainChunkSize::RECT_SIZE, |e, f| e / f as i32);
@@ -390,6 +490,89 @@ impl Client {
                     },
                     |pos, (r, g, b, a)| {
                         world_map_rgba[pos.y * map_size.x as usize + pos.x] =
+                            u32::from_le_bytes([r, g, b, a]);
+                    },
+                );
+                map_config.is_political = true;
+                map_config.generate(
+                    |pos| {
+                        sample_pos(
+                            &map_config,
+                            pos,
+                            &alt,
+                            &rgba,
+                            &map_size,
+                            &map_size_lg,
+                            max_height,
+                        )
+                    },
+                    |wpos| {
+                        let pos = wpos.map2(TerrainChunkSize::RECT_SIZE, |e, f| e / f as i32);
+                        rescale_height(if bounds_check(pos) {
+                            scale_height_big(alt[pos])
+                        } else {
+                            0.0
+                        })
+                    },
+                    |pos, (r, g, b, a)| {
+                        world_map_political[pos.y * map_size.x as usize + pos.x] =
+                            u32::from_le_bytes([r, g, b, a]);
+                    },
+                );
+
+                map_config.is_shaded = false;
+                map_config.rgba_alpha = 0.5;
+                map_config.is_political = false;
+                map_config.generate(
+                    |pos| {
+                        sample_pos(
+                            &map_config,
+                            pos,
+                            &alt,
+                            &rgba,
+                            &map_size,
+                            &map_size_lg,
+                            max_height,
+                        )
+                    },
+                    |wpos| {
+                        let pos = wpos.map2(TerrainChunkSize::RECT_SIZE, |e, f| e / f as i32);
+                        rescale_height(if bounds_check(pos) {
+                            scale_height_big(alt[pos])
+                        } else {
+                            0.0
+                        })
+                    },
+                    |pos, (r, g, b, a)| {
+                        world_map_rgba_half_alpha[pos.y * map_size.x as usize + pos.x] =
+                            u32::from_le_bytes([r, g, b, a]);
+                    },
+                );
+                // Generate topographic map
+                map_config.is_contours = true;
+                map_config.is_roads = true;
+                map_config.generate(
+                    |pos| {
+                        sample_pos(
+                            &map_config,
+                            pos,
+                            &alt,
+                            &rgba,
+                            &map_size,
+                            &map_size_lg,
+                            max_height,
+                        )
+                    },
+                    |wpos| {
+                        let pos = wpos.map2(TerrainChunkSize::RECT_SIZE, |e, f| e / f as i32);
+                        rescale_height(if bounds_check(pos) {
+                            scale_height_big(alt[pos])
+                        } else {
+                            0.0
+                        })
+                    },
+                    |pos, (r, g, b, a)| {
+                        world_map_topo[pos.y * map_size.x as usize + pos.x] =
                             u32::from_le_bytes([r, g, b, a]);
                     },
                 );
@@ -412,7 +595,16 @@ impl Client {
                 ping_stream.send(PingMsg::Ping)?;
                 let lod_base = rgba;
                 let lod_alt = alt;
-                let world_map_img = make_raw(&world_map_rgba)?;
+                let world_map_rgba_img = make_raw(&world_map_rgba)?;
+                let world_map_political_img = make_raw(&world_map_political)?;
+                let world_map_rgba_half_alpha_img = make_raw(&world_map_rgba_half_alpha)?;
+                let world_map_topo_img = make_raw(&world_map_topo)?;
+                let world_map_layers = vec![
+                    world_map_rgba_img,
+                    world_map_political_img,
+                    world_map_rgba_half_alpha_img,
+                    world_map_topo_img,
+                ];
                 let horizons = (west.0, west.1, east.0, east.1)
                     .into_par_iter()
                     .map(|(wa, wh, ea, eh)| u32::from_le_bytes([wa, wh, ea, eh]))
@@ -426,7 +618,7 @@ impl Client {
                     lod_base,
                     lod_alt,
                     Grid::from_raw(map_size.map(|e| e as i32), lod_horizon),
-                    (world_map_img, map_size, map_bounds),
+                    (world_map_layers, map_size, map_bounds),
                     world_map.sites,
                     recipe_book,
                     max_group_size,