+++ voxygen/src/hud/map.rs
@@ -31,7 +31,7 @@ widget_ids! {
         location_name,
         indicator,
         indicator_overlay,
-        grid,
+        map_layers[],
         map_title,
         qlog_title,
         zoom_slider,
@@ -40,6 +40,9 @@ widget_ids! {
         member_indicators[],
         member_height_indicators[],
         map_settings_align,
+        show_topo_map_img,
+        show_topo_map_box,
+        show_topo_map_text,
         show_towns_img,
         show_towns_box,
         show_towns_text,
@@ -71,7 +74,7 @@ const SHOW_ECONOMY: bool = false; // turn this display off (for 0.9) until we ha
 #[derive(WidgetCommon)]
 pub struct Map<'a> {
     client: &'a Client,
-    world_map: &'a (img_ids::Rotations, Vec2<u32>),
+    world_map: &'a (Vec<img_ids::Rotations>, Vec2<u32>),
     imgs: &'a Imgs,
     fonts: &'a Fonts,
     #[conrod(common_builder)]
@@ -88,7 +91,7 @@ impl<'a> Map<'a> {
         client: &'a Client,
         imgs: &'a Imgs,
         rot_imgs: &'a ImgsRot,
-        world_map: &'a (img_ids::Rotations, Vec2<u32>),
+        world_map: &'a (Vec<img_ids::Rotations>, Vec2<u32>),
         fonts: &'a Fonts,
         pulse: f32,
         localized_strings: &'a Localization,
@@ -123,6 +126,7 @@ pub enum Event {
     ShowDungeons(bool),
     ShowCaves(bool),
     ShowTrees(bool),
+    ShowTopoMap(bool),
     Close,
     RequestSiteInfo(SiteId),
 }
@@ -184,6 +188,7 @@ impl<'a> Widget for Map<'a> {
         let show_castles = self.global_state.settings.interface.map_show_castles;
         let show_caves = self.global_state.settings.interface.map_show_caves;
         let show_trees = self.global_state.settings.interface.map_show_trees;
+        let show_topo_map = self.global_state.settings.interface.map_show_topo_map;
         let mut events = Vec::new();
         let i18n = &self.localized_strings;
         // Tooltips
@@ -265,13 +270,25 @@ impl<'a> Widget for Map<'a> {
                 .color(TEXT_COLOR)
                 .set(state.ids.location_name, ui),
         }*/
+        // Map Layers
+        // It is assumed that there is at least one layer
+        if state.ids.map_layers.len() < self.world_map.0.len() {
+            state.update(|state| {
+                state
+                    .ids
+                    .map_layers
+                    .resize(self.world_map.0.len(), &mut ui.widget_id_generator())
+            });
+        }
+
         Image::new(self.imgs.map_frame_art)
             .mid_top_with_margin_on(state.ids.map_align, 5.0)
             .w_h(765.0, 765.0)
             .parent(state.ids.bg)
-            .set(state.ids.grid, ui);
-        // Map Image
-        let (world_map, worldsize) = self.world_map;
+            .set(state.ids.map_layers[0], ui);
+
+        // Map Size
+        let worldsize = self.world_map.1;
 
         // Coordinates
         let player_pos = self
@@ -292,7 +309,7 @@ impl<'a> Widget for Map<'a> {
         // Handle dragging
         let drag = self.global_state.settings.interface.map_drag;
         let dragged: Vec2<f64> = ui
-            .widget_input(state.ids.grid)
+            .widget_input(state.ids.map_layers[0])
             .drags()
             .left()
             .map(|drag| Vec2::<f64>::from(drag.delta_xy))
@@ -320,15 +337,30 @@ impl<'a> Widget for Map<'a> {
         {
             events.push(Event::Close);
         }
-        Image::new(world_map.none)
-            .mid_top_with_margin_on(state.ids.map_align, 10.0)
-            .w_h(map_size.x, map_size.y)
-            .parent(state.ids.bg)
-            .source_rectangle(rect_src)
-            .set(state.ids.grid, ui);
+
+        // Map Layer Images
+        for (index, layer) in self.world_map.0.iter().enumerate() {
+            if index == 0 {
+                Image::new(layer.none)
+                    .mid_top_with_margin_on(state.ids.map_align, 10.0)
+                    .w_h(map_size.x, map_size.y)
+                    .parent(state.ids.bg)
+                    .source_rectangle(rect_src)
+                    .set(state.ids.map_layers[index], ui);
+            } else if show_topo_map {
+                Image::new(layer.none)
+                    .mid_top_with_margin_on(state.ids.map_align, 10.0)
+                    .w_h(map_size.x, map_size.y)
+                    .parent(state.ids.bg)
+                    .source_rectangle(rect_src)
+                    .graphics_for(state.ids.map_layers[0])
+                    .set(state.ids.map_layers[index], ui);
+            }
+        }
+
         // Handle zooming with the mousewheel
         let scrolled: f64 = ui
-            .widget_input(state.ids.grid)
+            .widget_input(state.ids.map_layers[0])
             .scrolls()
             .map(|scroll| scroll.y)
             .sum();
@@ -342,9 +374,44 @@ impl<'a> Widget for Map<'a> {
             .top_right_with_margins_on(state.ids.frame, 55.0, 10.0)
             .set(state.ids.map_settings_align, ui);
         // Checkboxes
+        // Show topographic map
+        Image::new(self.imgs.map_topo)
+            .top_left_with_margins_on(state.ids.map_settings_align, 5.0, 5.0)
+            .w_h(20.0, 20.0)
+            .set(state.ids.show_topo_map_img, ui);
+        if Button::image(if show_topo_map {
+            self.imgs.checkbox_checked
+        } else {
+            self.imgs.checkbox
+        })
+        .w_h(18.0, 18.0)
+        .hover_image(if show_topo_map {
+            self.imgs.checkbox_checked_mo
+        } else {
+            self.imgs.checkbox_mo
+        })
+        .press_image(if show_topo_map {
+            self.imgs.checkbox_checked
+        } else {
+            self.imgs.checkbox_press
+        })
+        .right_from(state.ids.show_topo_map_img, 10.0)
+        .set(state.ids.show_topo_map_box, ui)
+        .was_clicked()
+        {
+            events.push(Event::ShowTopoMap(!show_topo_map));
+        }
+        Text::new(i18n.get("hud.map.topo_map"))
+            .right_from(state.ids.show_topo_map_box, 10.0)
+            .font_size(self.fonts.cyri.scale(14))
+            .font_id(self.fonts.cyri.conrod_id)
+            .graphics_for(state.ids.show_topo_map_box)
+            .color(TEXT_COLOR)
+            .set(state.ids.show_topo_map_text, ui);
+
         // Show difficulties
         Image::new(self.imgs.map_dif_6)
-            .top_left_with_margins_on(state.ids.map_settings_align, 5.0, 5.0)
+            .down_from(state.ids.show_topo_map_img, 10.0)
             .w_h(20.0, 20.0)
             .set(state.ids.show_difficulty_img, ui);
         if Button::image(if show_difficulty {
@@ -576,6 +643,7 @@ impl<'a> Widget for Map<'a> {
             // Convert to relative pixel coordinates from the center of the map
             // Accounting for zooming
             let rpos = rfpos.map2(map_size, |e, sz| e * sz as f32 * zoom as f32);
+            let rside = zoom * 6.0;
 
             if rpos
                 .map2(map_size, |e, sz| e.abs() > sz as f32 / 2.0)
@@ -610,11 +678,11 @@ impl<'a> Widget for Map<'a> {
                 SiteKind::Tree => self.imgs.mmap_site_tree,
             })
             .x_y_position_relative_to(
-                state.ids.grid,
+                state.ids.map_layers[0],
                 position::Relative::Scalar(rpos.x as f64),
                 position::Relative::Scalar(rpos.y as f64),
             )
-            .w_h(20.0 * 1.2, 20.0 * 1.2)
+            .w_h(rside * 1.2, rside * 1.2)
             .hover_image(match &site.kind {
                 SiteKind::Town => self.imgs.mmap_site_town_hover,
                 SiteKind::Dungeon { .. } => self.imgs.mmap_site_dungeon_hover,
@@ -673,16 +741,16 @@ impl<'a> Widget for Map<'a> {
                     _ => self.imgs.nothing,
                 })
                 .mid_top_with_margin_on(state.ids.mmap_site_icons[i], match difficulty {
-                    5 => -12.0 * size,
-                    _ => -4.0 * size,
+                    5 => -2.0 * zoom * size,
+                    _ => -1.0 * zoom * size,
                 })
                 .w(match difficulty {
-                    5 => 12.0 * size,
-                    _ => 4.0 * size * difficulty as f64,
+                    5 => 2.0 * zoom * size,
+                    _ => 1.0 * zoom * size * difficulty as f64,
                 })
                 .h(match difficulty {
-                    5 => 12.0 * size,
-                    _ => 4.0 * size,
+                    5 => 2.0 * size * zoom,
+                    _ => 1.0 * zoom * size,
                 })
                 .color(Some(match difficulty {
                     0 => QUALITY_LOW,
@@ -785,7 +853,7 @@ impl<'a> Widget for Map<'a> {
                     _ => self.imgs.indicator_group,
                 })
                 .x_y_position_relative_to(
-                    state.ids.grid,
+                    state.ids.map_layers[0],
                     position::Relative::Scalar(rpos.x as f64),
                     position::Relative::Scalar(rpos.y as f64),
                 )
@@ -819,7 +887,7 @@ impl<'a> Widget for Map<'a> {
         {
             Image::new(self.rot_imgs.indicator_mmap_small.target_north)
                 .x_y_position_relative_to(
-                    state.ids.grid,
+                    state.ids.map_layers[0],
                     position::Relative::Scalar(rpos.x as f64),
                     position::Relative::Scalar(rpos.y as f64),
                 )
@@ -838,7 +906,7 @@ impl<'a> Widget for Map<'a> {
         };
         if Button::image(self.imgs.button)
             .w_h(92.0, icon_size.y)
-            .mid_bottom_with_margin_on(state.ids.grid, -36.0)
+            .mid_bottom_with_margin_on(state.ids.map_layers[0], -36.0)
             .hover_image(if recenter {
                 self.imgs.button_hover
             } else {
@@ -870,7 +938,7 @@ impl<'a> Widget for Map<'a> {
         };
 
         Image::new(self.imgs.m_move_ico)
-            .bottom_left_with_margins_on(state.ids.grid, -36.0, 0.0)
+            .bottom_left_with_margins_on(state.ids.map_layers[0], -36.0, 0.0)
             .w_h(icon_size.x, icon_size.y)
             .color(Some(UI_HIGHLIGHT_0))
             .set(state.ids.drag_ico, ui);
@@ -878,7 +946,7 @@ impl<'a> Widget for Map<'a> {
             .right_from(state.ids.drag_ico, 5.0)
             .font_size(self.fonts.cyri.scale(14))
             .font_id(self.fonts.cyri.conrod_id)
-            .graphics_for(state.ids.grid)
+            .graphics_for(state.ids.map_layers[0])
             .color(TEXT_COLOR)
             .set(state.ids.drag_txt, ui);
         Image::new(self.imgs.m_scroll_ico)
@@ -890,7 +958,7 @@ impl<'a> Widget for Map<'a> {
             .right_from(state.ids.zoom_ico, 5.0)
             .font_size(self.fonts.cyri.scale(14))
             .font_id(self.fonts.cyri.conrod_id)
-            .graphics_for(state.ids.grid)
+            .graphics_for(state.ids.map_layers[0])
             .color(TEXT_COLOR)
             .set(state.ids.zoom_txt, ui);